pipeline/pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kaggle>=1.8.3",
    "openai>=2.15.0",
    "polars>=1.37.1",
    "pydantic>=2.12.5",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "tenacity>=9.1.2",
    "typer>=0.21.1",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "ipykernel>=7.1.0",
    "pip>=25.3",
    "tiktoken>=0.12.0",
]

[tool.uv]
package = true

[project.scripts]
pipeline = "pipeline.cli:app"
